declare @effects_data as a tracked var
Effect object data will be stored as hashes to replace
the $fx->{applied} and $fx->{params} data structures currently used.

The class must be munged, so that on restore, the
methods come from the new class with accessors based on the
hash structure, rather than the previous, with accessors
based on $fx->{applied} and $fx->{params}

On save, don't save $fx->{applied} and $fx->{params}
if saving @effects_data.